Initial work on mix.lock

+++ lib/mix/lib/mix/scm/git.ex
@@ -24,13 +24,32 @@ defmodule Mix.SCM.Git do
         maybe_error System.cmd("git fetch --force --quiet --tags")
       end
 
-      ref = opts[:ref] || "master"
-      maybe_error System.cmd("git reset --hard --quiet #{ref}")
+      if available?(".") do
+        ref = opts[:ref] || "master"
+        maybe_error System.cmd("git reset --hard --quiet #{ref}")
 
-      if opts[:submodules] do
-        maybe_error System.cmd("git submodule update --init --recursive")
+        if opts[:submodules] do
+          maybe_error System.cmd("git submodule update --init --recursive")
+        end
+
+        check_rev System.cmd('git rev-parse --verify --quiet #{ref}')
       end
     end
+  end
+
+  defp check_rev([]),   do: nil
+  defp check_rev(list), do: check_rev(list, [])
+
+  defp check_rev([h|t], acc) when h in ?a..?f or h in ?0..?9 do
+    check_rev(t, [h|acc])
+  end
+
+  defp check_rev(fin, acc) when fin == [?\n] or fin == [] do
+    List.reverse(acc) /> list_to_binary
+  end
+
+  defp check_rev(_, _) do
+    nil
   end
 
   defp maybe_error(""),    do: :ok
